WIP: RelationProof has "a" and "b" profile, peer_id(my_id) always returns the peer.
The serialization of the RelationProof is always "a" first. Since it's in alphabetic
order, the initiator cannot be found out.

TODO: We need to have the relation type pairs because we need to convert when creating
a RelationProof of a RelationHalfProof. Examples:
"a_is_home_of_b" - "b_is_home_of_a" (asymmetric)
"friend" - "friend" (symmetric)

Alternate solution with better performance, but more headaches:
 * call relationproof.set_my_id(my_profile_id) once (it needs to be mut!)
 * wrap relationproof to a structure that knows which profile is the peer (seems complicated)